nhanh test master 11

+++ src/test/java/TestInterFace.java
@@ -6,5 +6,5 @@ public interface TestInterFace extends TestInterFace2,TestInterFace3 {
 
       void Tc_01();
 
-    int Tc_02();
+    long Tc_02();
 }
